+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## HEAD
 
   * feature: change cookie placement locator mechanism to use injections and stacktrace.js instead of Debugger.scriptParsed for improved accuracy
+  * feature: also track HTTP cookies 
 
 ## 0.1.0 / 2019-04-17